tweaks towards clarity for coloring challenges
lindsey (via RC Zulip) pointed out how confused she was about what was a 'challenge' or a 'coloring challenge'.  A bigger audit with fixes is needed to resolve this ambiguity.  Small tweaks are address in this commit.

The inconsistency in which shapes correspond to the coloring challenges is confusing
“I was a bit confused by some of the challenge problems. On this page, for instance, the problem is "Challenge: Which dihedral group does each uncolored shape belong to?", and I didn't know if that was referring to the shapes above or on the right or what. Is there a missing illustration below? If it's supposed to be referring to the shapes on the right, then the problem seems too easy, since they're already arranged into rows by group.
You’re right that the shapes are already lined up in rows by groups and that this is easy once you see that.  I thought some people might not immediately see that and wanted this as a checkpoint.  You are so right that this is unclear!  You also helped me realize that the word “uncolored” only adds confusion.  I replaced it with “illustrated” in the hope that would guide people to the “illustrations” on the right.  I wonder if this helps at all

For some of the other challenges, like on this page, the challenge is about the shapes on the right. Actually, I just realized that maybe that's what "Coloring Challenge" means, as opposed to just "Challenge". Maybe this could be explained up front.
As a tweak towards clarity I updated this to read “Coloring Challenge: Color the C4 shapes to make them all C2 shapes.” and put the caption of “C4 shapes” below the illustrations on the right.

But then, this page has a "Coloring Challenge" but also has shapes below. I'm not sure if the challenge is supposed to be about those, or about the ones on the right.
With all these ambiguities, I started to wish that the book had hidden solutions to the challenge problems that could be revealed by clicking. That might be a lot of effort to add, but I bet people would be interested in trying to check their work.
I really do wish I had created a solution sheet of some sort, I think that will be the next iteration of this project!

For example, on this page, it's called a "coloring challenge", but it seems to only have to do with the content directly below it, not the content on the right.
I don't think I have a great idea for resolving the confusion everywhere, but one thing that might help would be to audit where you're saying things like "the triangles" or "the shapes" in the challenges, and instead say, e.g., "the three triangles below" or "the sixteen shapes on the right", to make it obvious what you're talking about.
Also, I'm still not sure if a "coloring challenge" is meant to be a distinct thing from a "challenge".
Thanks again for pointing this out.  I made another tweak with the intention of more clarity for this page: http://www.coloring-book.co/?pageName=cn-15&pageNumber=21
The coloring challenge was intended for both the below shapes as well as the shapes on the right.
Now the coloring challenge is about C4 and C6 shapes, and is only C4 and C6 shapes directly below and to the right, with the caption on the right indicating so.