created sample of consolidated introduction and rotations sections to show to the MIT Press
see summary of changes in https://docs.google.com/document/d/1W5S6lf-o9raqnQncRHoRrINBn0R-JP3AJG7--rFrW0I/edit#
An overview of the new proposal’s change in structure:

The content is structured into denser sections.
Each section covers a few mathematical concepts and intersperses denser text with explanatory illustrations and “check ins”.
Each section is followed by a set of exercises that set the reader to task with “challenges” to engage with the concepts presented and illustrated symmetries.  The reviewers pointed out that the coloring challenges are innovative and make this book shine, but that they may limit the ability to reuse this book.  For this reason, the pages with challenge exercises are separated from the main text so that they can separately be photocopied or otherwise provided and then colored in.  This will allow the reader to make mistakes and allow the book to more easily be reused and serve a classroom.
In the patterns part of the book, some of the challenges can be presented on pages opposite of the illustrations so that the illustrations can take up their entire page with a full bleed.  (See the attached example).
As suggested by the reviewers, the restructured version of the book has more challenges and more deeply engages the reader in the concepts covered after each section.
Sample solutions to challenges are to be provided at the end of the book.
The denser text sections have been filled in with more content, to address the reviewers requests for more elaboration of the concepts and that the author’s voice better shine through in the explanations.
Despite the additional text, the restructured version is a shorter book due to the consolidation of pages.

The reviewers recommended adding additional content, so the new version will also include:
Why the only possible rotations for wallpaper groups are ½, ⅓, ¼, ⅙ turns.
Information on the naming schemes of the groups.

The reviewers also requested more coloring challenges that use color to engage with the concept of subgroups. “The use of coloring to demonstrate what a subgroup is works really well!!” (reviewer 1).  The new version will include more of these challenges.

They also pointed out that the reader may have more questions and interest in the subject inspired by reading the book, and that the book should point them in the right next directions to further explore these interests.  The conclusion of the proposed new version will refer the reader to where they can go next.

The proposed new version of the book can still be complemented by an interactive digital version that brings all of the illustrations to life in animation (see the original at http://coloring-book.co).  This blend of interactive digital and paper mediums makes this book about math unique.

For the paper version, please note that while it invites color, it does not require color printing.  The colors in the book are as demonstrative in grayscale.  Where the figures have multiple colors, the multiple colors have differing textures so that the difference can still be seen when printed in black and white.